LayoutTests/http/tests/security/video-poster-cross-origin-crash-expected.txt

CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Cannot load image http://localhost:8080/misc/resources/compass.jpg due to access control checks.
>>>
